requirements.txt
azure-identity==1.13.0
langchain==0.0.236
openai[datalib]==0.27.8
GitPython==3.1.40
azure-devops==7.1.0b3
PyGithub